.github/actions/monorepo-config/requirements.txt
attrs==21.2.0
click==8.0.1
colorama==0.4.4
commonmark==0.9.1
docopt==0.6.2
hypothesis==6.14.0
iniconfig==1.1.1
json5==0.9.5
packaging==20.9
pluggy==0.13.1
py==1.10.0
Pygments==2.9.0
pyparsing==2.4.7
pystache==0.5.4
pytest==6.2.4
pytest-watch==4.2.0
PyYAML==5.4.1
rich==10.2.2
sortedcontainers==2.4.0
toml==0.10.2
toolz==0.11.1
watchdog==2.1.2
